app/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
joblib==1.5.2
lxml==6.0.1
numpy==2.3.3
pandas==2.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0

# AI / NLP models
torch>=2.0.0
transformers>=4.44.0
sentence-transformers>=3.0.1
